+++ GeneratorDuty/Commands/Duty/ForceCommand.cs
@@ -1,16 +1,13 @@
 using GeneratorDuty.Common;
 using GeneratorDuty.CustomRights;
-using GeneratorDuty.Database;
 using GeneratorDuty.Extensions;
 using GeneratorDuty.Models;
 using GeneratorDuty.Repository;
-using GeneratorDuty.Services;
-using Microsoft.EntityFrameworkCore;
 using Telegram.Bot;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.ReplyMarkups;
 
-namespace GeneratorDuty.Commands;
+namespace GeneratorDuty.Commands.Duty;
 
 public class ForceCommand(DutyRepository repository) : BaseCommand
 {
