+++ GeneratorDuty/Commands/Schedule/AutoCommand.cs
@@ -1,12 +1,10 @@
 using GeneratorDuty.Common;
-using GeneratorDuty.Database;
 using GeneratorDuty.Extensions;
 using GeneratorDuty.Repository;
-using Microsoft.EntityFrameworkCore;
 using Telegram.Bot;
 using Telegram.Bot.Types;
 
-namespace GeneratorDuty.Commands;
+namespace GeneratorDuty.Commands.Schedule;
 
 public class AutoCommand(DutyRepository repository) : BaseCommand
 {
