+++ GeneratorDuty/Commands/Schedule/SAutoCommand.cs
@@ -1,14 +1,12 @@
 using ClientSamgk.Enums;
 using GeneratorDuty.Common;
-using GeneratorDuty.Database;
 using GeneratorDuty.Extensions;
 using GeneratorDuty.Models;
 using GeneratorDuty.Repository;
-using Microsoft.EntityFrameworkCore;
 using Telegram.Bot;
 using Telegram.Bot.Types;
 
-namespace GeneratorDuty.Commands;
+namespace GeneratorDuty.Commands.Schedule;
 
 public class SAutoCommand(DutyRepository repository) : BaseCommand
 {
