+++ GeneratorDuty/Telegrams/UpdateHandle.cs
@@ -3,6 +3,9 @@ using GeneratorDuty.CallBackKeyboards;
 using GeneratorDuty.CallBackKeyboards.Duty;
 using GeneratorDuty.CallBackKeyboards.Schedule;
 using GeneratorDuty.Commands;
+using GeneratorDuty.Commands.Duty;
+using GeneratorDuty.Commands.Schedule;
+using GeneratorDuty.Commands.Service;
 using GeneratorDuty.Common;
 using GeneratorDuty.Database;
 using GeneratorDuty.Extensions;
